feat: support top level await
