+++ examples/top-level-await/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "root": true,
+  "extends": "../../.eslintrc.base.json",
+  "parserOptions": {
+    "project": ["./examples/top-level-await/tsconfig.json"]
+  },
+  "rules": {}
+}
