+++ examples/top-level-await/src/data.js
@@ -1,0 +1,6 @@
+export default [
+  'data1',
+  'data2',
+  'data3',
+  'data4',
+];
