+++ examples/top-level-await/src/dep1.js
@@ -1,0 +1,1 @@
+export { default } from './dep2.js';
