+++ examples/top-level-await/src/index.js
@@ -1,0 +1,3 @@
+import dep1 from './dep1.js';
+
+console.log(dep1);
